refactor(ocm): simplify token path coupling
