+++ internal/http/services/opencloudmesh/ocmd/ocm.go
@@ -22,7 +22,6 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/cs3org/reva/v3/internal/http/services/opencloudmesh/contract"
 	"github.com/cs3org/reva/v3/pkg/appctx"
 	"github.com/cs3org/reva/v3/pkg/rhttp/global"
 	"github.com/cs3org/reva/v3/pkg/sharedconf"
@@ -98,10 +97,10 @@ func (s *svc) routerInit() error {
 		return err
 	}
 
-	s.router.Post(contract.SharesPath, sharesHandler.CreateShare)
-	s.router.Post(contract.InviteAcceptedPath, invitesHandler.AcceptInvite)
-	s.router.Post(contract.NotificationsPath, notifHandler.Notifications)
-	s.router.Post(contract.TokenPath, tokenHandler.ExchangeToken)
+	s.router.Post(sharesPath, sharesHandler.CreateShare)
+	s.router.Post(inviteAcceptedPath, invitesHandler.AcceptInvite)
+	s.router.Post(notificationsPath, notifHandler.Notifications)
+	s.router.Post(tokenPath, tokenHandler.ExchangeToken)
 	return nil
 }
 
@@ -117,7 +116,7 @@ func (s *svc) Prefix() string {
 func (s *svc) Unprotected() []string {
 	// These OCM ingress routes authenticate at the protocol layer, so they stay
 	// reachable without the outer auth middleware.
-	return []string{contract.InviteAcceptedPath, contract.SharesPath, contract.NotificationsPath, contract.TokenPath}
+	return []string{inviteAcceptedPath, sharesPath, notificationsPath, tokenPath}
 }
 
 func (s *svc) Handler() http.Handler {
