+++ internal/http/services/wellknown/ocm.go
@@ -24,7 +24,6 @@ import (
 	"net/url"
 	"path/filepath"
 
-	"github.com/cs3org/reva/v3/internal/http/services/opencloudmesh/contract"
 	"github.com/cs3org/reva/v3/pkg/appctx"
 )
 
@@ -146,14 +145,15 @@ func (h *wkocmHandler) init(c *OcmProviderConfig) {
 	d.Capabilities = []string{"invites", "webdav-uri", "protocol-object", "invite-wayf"}
 	d.InviteAcceptDialog, _ = url.JoinPath(c.Endpoint, c.InviteAcceptDialog)
 	if c.EnableCodeFlow {
-		d.TokenEndPoint, _ = tokenEndpoint(c.Endpoint, c.OCMPrefix)
+		d.TokenEndPoint, _ = TokenEndpoint(c.Endpoint, c.OCMPrefix)
 		d.Capabilities = append(d.Capabilities, "exchange-token")
 	}
 	h.data = d
 }
 
-func tokenEndpoint(baseURL, prefix string) (string, error) {
-	return contract.TokenEndpoint(baseURL, prefix)
+// TokenEndpoint builds the advertised code-flow token endpoint for OCM discovery.
+func TokenEndpoint(baseURL, prefix string) (string, error) {
+	return url.JoinPath(baseURL, prefix, "token")
 }
 
 // Ocm handles the OCM discovery endpoint specified in
